+++ eclipsetools/preprocessing/masking.py
@@ -11,7 +11,7 @@ def hann_window_mask(shape: tuple) -> np.ndarray:
 def circle_mask(shape: np.ndarray,
                 mask_center: tuple,
                 mask_radius: float) -> np.ndarray:
-    mask = np.zeros(shape)
+    mask = np.zeros(shape, dtype=np.float32)
     mask = 1 - cv2.circle(
         img=mask,
         center=(int(mask_center[1]), int(mask_center[0])),
